chore: dedupe gossipsub benchmark runBenchmark helper
Extract the per-file runBenchmark helper (identical except protobuf's
runsFactor, now defaulted to 1) into test/utils/benchmark.ts.
