+++ packages/gossipsub/test/benchmark/asyncIterable.test.ts
@@ -1,8 +1,7 @@
 import { abortableSource } from 'abortable-iterator'
-// @ts-expect-error no types
-import Benchmark from 'benchmark'
 import all from 'it-all'
 import { pipe } from 'it-pipe'
+import { runBenchmark } from '../utils/benchmark.js'
 
 describe('abortableSource cost', function () {
   const n = 10000
@@ -102,27 +101,3 @@ describe('pipe extra iterables cost', function () {
     })
   })
 })
-
-async function runBenchmark (name: string, fn: () => void | Promise<void>): Promise<void> {
-  await new Promise<void>((resolve, reject) => {
-    new Benchmark(name, {
-      defer: true,
-      initCount: 1,
-      maxTime: 1,
-      minSamples: 1,
-      minTime: 0.1,
-      fn (deferred: { resolve(): void }) {
-        Promise.resolve()
-          .then(fn)
-          .then(() => { deferred.resolve() })
-          .catch(reject)
-      }
-    })
-      .on('complete', function (this: { hz: number, count: number }) {
-        process.stdout.write(`    ${name}: ${this.hz.toFixed(4)} ops/s ${(1000 / this.hz).toFixed(6)} ms/op ${this.count} runs\n`)
-        resolve()
-      })
-      .on('error', reject)
-      .run({ async: true })
-  })
-}
