+++ packages/gossipsub/test/benchmark/time-cache.test.ts
@@ -1,8 +1,7 @@
 // @ts-expect-error no types
-import Benchmark from 'benchmark'
-// @ts-expect-error no types
 import TimeCache from 'time-cache'
 import { SimpleTimeCache } from '../../src/utils/time-cache.js'
+import { runBenchmark } from '../utils/benchmark.js'
 
 // TODO: errors with "Error: root suite not found"
 describe('npm TimeCache vs SimpleTimeCache', () => {
@@ -24,27 +23,3 @@ describe('npm TimeCache vs SimpleTimeCache', () => {
     })
   }
 })
-
-async function runBenchmark (name: string, fn: () => void | Promise<void>): Promise<void> {
-  await new Promise<void>((resolve, reject) => {
-    new Benchmark(name, {
-      defer: true,
-      initCount: 1,
-      maxTime: 1,
-      minSamples: 1,
-      minTime: 0.1,
-      fn (deferred: { resolve(): void }) {
-        Promise.resolve()
-          .then(fn)
-          .then(() => { deferred.resolve() })
-          .catch(reject)
-      }
-    })
-      .on('complete', function (this: { hz: number, count: number }) {
-        process.stdout.write(`    ${name}: ${this.hz.toFixed(4)} ops/s ${(1000 / this.hz).toFixed(6)} ms/op ${this.count} runs\n`)
-        resolve()
-      })
-      .on('error', reject)
-      .run({ async: true })
-  })
-}
